Refactor language handling to use shared `i18n` package
- Replaced `lang` package with `i18n` for language parsing, defaults, and context handling.
- Updated middleware and user service to adopt the new shared package.
- Moved and renamed `language.go` from `domain/lang` to `shared/i18n`.
